# user-service/docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.BaseResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  controllers.CreateUserRequest:
    properties:
      email:
        type: string
    type: object
  controllers.GetBalanceResponse:
    properties:
      balance:
        type: number
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a user service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: User Service API
  version: "1.0"
paths:
  /balance/{email}:
    get:
      description: Get the balance of a user by email
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.GetBalanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.BaseResponse'
      summary: Get user balance
      tags:
      - users
  /createUser:
    post:
      consumes:
      - application/json
      description: Create a new user with the provided email
      parameters:
      - description: User email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.BaseResponse'
      summary: Create a new user
      tags:
      - users
swagger: "2.0"